moved files & added useRouter
